Split Program.cs to add modularity
